+++ packages/core-data/src/style.scss
@@ -1,0 +1,1 @@
+@import "./components/entities-saved-states/style.scss";
