+++ packages/core-data/src/utils/get-template-part-icon.js
