+++ packages/core-data/tsconfig.json
@@ -7,6 +7,7 @@
 	},
 	"references": [
 		{ "path": "../api-fetch" },
+		{ "path": "../components" },
 		{ "path": "../compose" },
 		{ "path": "../block-editor" },
 		{ "path": "../data" },
@@ -14,7 +15,9 @@
 		{ "path": "../element" },
 		{ "path": "../html-entities" },
 		{ "path": "../i18n" },
+		{ "path": "../icons" },
 		{ "path": "../is-shallow-equal" },
+		{ "path": "../notices" },
 		{ "path": "../private-apis" },
 		{ "path": "../rich-text" },
 		{ "path": "../sync" },
