+++ packages/editor/src/style.scss
@@ -10,7 +10,6 @@
 @import "./components/document-tools/style.scss";
 @import "./components/editor-interface/style.scss";
 @import "./components/editor-notices/style.scss";
-@import "./components/entities-saved-states/style.scss";
 @import "./components/error-boundary/style.scss";
 @import "./components/header/style.scss";
 @import "./components/inserter-sidebar/style.scss";
